pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.49.1",
    "babel>=2.18.0",
    "boto3>=1.42.96",
    "fastapi>=0.136.0",
    "faster-whisper>=1.2.1",
    "google-genai>=1.73.1",
    "pydantic-settings>=2.14.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.45.0",
]
